Cleanup types and handle pagination qs
