--- types.ts
@@ -1,8 +1,0 @@
-export type ApiSuccess<T> = { data: T; pageToken?: string };
-
-export interface ApiError extends Error {
-  info?: string;
-  status: number;
-}
-
-export type ApiResponse<T = any> = ApiSuccess<T> | { error: ApiError };
