+++ types/base.ts
@@ -1,10 +1,16 @@
-export type ApiError = {
-  code?: string;
-  message: string;
-  values: { [key: string]: string };
-};
+export type ApiSuccess<T> = { data: T; pageToken?: string };
 
-export type ApiResponse<T> = {
-  data: T | null;
-  error: ApiError | null;
-};
+export interface ApiError extends Error {
+  info?: string;
+  status: number;
+}
+
+export type ApiResponse<T = any> = ApiSuccess<T> | { error: ApiError };
+
+export type PaginateApiParams = (
+  | {
+      offset: string;
+      limit: string;
+    }
+  | { pageOffset: string; pageLimit: string }
+) & { pageToken?: string };
